consistent casing, removed namespace
snake case, camel case, and camelCase. I've used camelCase across, but generally, for query paths: kebab-case and snake_case for query parameters. camelCase for json payloads. 

https://github.com/saifaustcse/api-best-practices for more info.

namespaces need a formal definition and place in the model to make it to the swagger api. As of now, it is a consideration but not formalized so removed. 

will be looking at more later.